+++ src/main/kotlin/me/ffl/intellijDirectoryTests/Documentation.kt
@@ -4,18 +4,23 @@ import io.kotest.matchers.shouldBe
 import me.ffl.intellijDirectoryTests.MarkupFile.Companion.findCarets
 import kotlin.io.path.div
 import kotlin.io.path.readText
+import kotlin.io.path.writeText
 
 val documentationExecutor: KotestExecutor = {
     val beforeDir = testDataPath / "project"
     val projectFiles = beforeDir.loadProject()
     projectFiles.findCarets().forEach { caret ->
-        val doc = caret.file.getDocumentationAt(caret.offset)
+        val doc = caret.file.getDocumentationAt(caret.offset)?.trimEnd()
         if (doc == null) {
             if (caret.name == null) fail("No documentation found at nameless caret")
             else fail("No documentation found at caret ${caret.name}")
         } else {
-            val expected = (testDataPath / "result.html").readText()
-            doc.trimEnd() shouldBe expected.trimEnd()
+            val resultFile = testDataPath / "result.html"
+            val expected = resultFile.readText().trimEnd()
+            if (doc != expected && config.overrideDocumentationOutput) {
+                resultFile.writeText(doc)
+            }
+            doc shouldBe expected
         }
     }
 }